requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
websockets==12.0

# NetCDF and geospatial
xarray==2023.12.0
netCDF4==1.6.5
h5netcdf==1.3.0
rioxarray==0.15.0
rasterio==1.3.9

# AWS S3 for Mapbox uploads
boto3==1.29.7

# Data processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4

# Visualization
matplotlib==3.7.2
pillow==10.1.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1

# CLI support
click==8.1.7
rich==13.7.0